# Permit partial iteration in On Demand