--- a/include/simdjson/generic/ondemand/array.h
+++ b/include/simdjson/generic/ondemand/array.h
@@ -28,13 +28,13 @@ public:
    *
    * Part of the std::iterable interface.
    */
-  simdjson_really_inline array_iterator<array> begin() & noexcept;
+  simdjson_really_inline array_iterator begin() noexcept;
   /**
    * Sentinel representing the end of the array.
    *
    * Part of the std::iterable interface.
    */
-  simdjson_really_inline array_iterator<array> end() & noexcept;
+  simdjson_really_inline array_iterator end() noexcept;
 
 protected:
   /**
@@ -44,8 +44,15 @@ protected:
    *        resulting array.
    * @error INCORRECT_TYPE if the iterator is not at [.
    */
-  static simdjson_really_inline simdjson_result<array> start(json_iterator_ref &&iter) noexcept;
-  static simdjson_really_inline simdjson_result<array> start(const uint8_t *json, json_iterator_ref &&iter) noexcept;
+  static simdjson_really_inline simdjson_result<array> start(value_iterator &iter) noexcept;
+  /**
+   * Begin array iteration.
+   *
+   * @param iter The iterator. Must be where the initial [ is expected. Will be *moved* into the
+   *        resulting array.
+   * @error INCORRECT_TYPE if the iterator is not at [.
+   */
+  static simdjson_really_inline simdjson_result<array> try_start(value_iterator &iter) noexcept;
   /**
    * Begin array iteration.
    *
@@ -54,7 +61,7 @@ protected:
    *
    * @param iter The iterator. Must be after the initial [. Will be *moved* into the resulting array.
    */
-  static simdjson_really_inline array started(json_iterator_ref &&iter) noexcept;
+  static simdjson_really_inline array started(value_iterator &iter) noexcept;
 
   /**
    * Create an array at the given Internal array creation. Call array::start() or array::started() instead of this.
@@ -63,30 +70,19 @@ protected:
    *        == true, or past the [] with is_alive() == false if the array is empty. Will be *moved*
    *        into the resulting array.
    */
-  simdjson_really_inline array(json_iterator_ref &&iter) noexcept;
-
-  //
-  // For array_iterator
-  //
-  simdjson_really_inline json_iterator &get_iterator() noexcept;
-  simdjson_really_inline json_iterator_ref borrow_iterator_child() noexcept;
-  simdjson_really_inline bool is_iterator_alive() const noexcept;
-  simdjson_really_inline void start_iterator() noexcept;
-  simdjson_really_inline void finish_iterator() noexcept;
-  simdjson_really_inline void abandon_iterator() noexcept;
-  simdjson_warn_unused simdjson_really_inline error_code finish_iterator_child() noexcept;
+  simdjson_really_inline array(const value_iterator &iter) noexcept;
 
   /**
    * Iterator marking current position.
    *
    * iter.is_alive() == false indicates iteration is complete.
    */
-  json_iterator_ref iter{};
+  value_iterator iter{};
 
   friend class value;
   friend struct simdjson_result<value>;
   friend struct simdjson_result<array>;
-  friend class array_iterator<array>;
+  friend class array_iterator;
 };
 
 } // namespace ondemand
@@ -105,8 +101,8 @@ public:
   simdjson_really_inline simdjson_result(simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array> &&a) noexcept = default;
   simdjson_really_inline ~simdjson_result() noexcept = default; ///< @private
 
-  simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<SIMDJSON_IMPLEMENTATION::ondemand::array>> begin() & noexcept;
-  simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<SIMDJSON_IMPLEMENTATION::ondemand::array>> end() & noexcept;
+  simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator> begin() noexcept;
+  simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator> end() noexcept;
 };
 
 } // namespace simdjson