--- a/include/simdjson/generic/ondemand/array_iterator-inl.h
+++ b/include/simdjson/generic/ondemand/array_iterator-inl.h
@@ -2,43 +2,27 @@ namespace simdjson {
 namespace SIMDJSON_IMPLEMENTATION {
 namespace ondemand {
 
-template<typename T>
-simdjson_really_inline array_iterator<T>::array_iterator(T &_iter) noexcept : iter{&_iter} {}
+simdjson_really_inline array_iterator::array_iterator(const value_iterator &_iter) noexcept
+  : iter{_iter}
+{}
 
-template<typename T>
-simdjson_really_inline simdjson_result<array_iterator<T>> array_iterator<T>::start(T &iter, const uint8_t *json) noexcept {
-  bool has_value;
-  SIMDJSON_TRY( iter.get_iterator().start_array(json).get(has_value) );
-  if (has_value) { iter.start_iterator(); } else { iter.abandon_iterator(); }
-  return array_iterator<T>(iter);
+simdjson_really_inline simdjson_result<value> array_iterator::operator*() noexcept {
+  if (iter.error()) { iter.abandon(); return iter.error(); }
+  return value(iter.child());
 }
-template<typename T>
-simdjson_really_inline simdjson_result<value> array_iterator<T>::operator*() noexcept {
-  error_code error = iter->get_iterator().error();
-  if (error) { iter->abandon_iterator(); return error; }
-  return value::start(iter->borrow_iterator_child());
-}
-template<typename T>
-simdjson_really_inline bool array_iterator<T>::operator==(const array_iterator<T> &other) const noexcept {
+simdjson_really_inline bool array_iterator::operator==(const array_iterator &other) const noexcept {
   return !(*this != other);
 }
-template<typename T>
-simdjson_really_inline bool array_iterator<T>::operator!=(const array_iterator<T> &) const noexcept {
-  return iter->is_iterator_alive();
+simdjson_really_inline bool array_iterator::operator!=(const array_iterator &) const noexcept {
+  return iter.is_open();
 }
-template<typename T>
-simdjson_really_inline array_iterator<T> &array_iterator<T>::operator++() noexcept {
-  // TODO this is a safety rail ... users should exit loops as soon as they receive an error.
-  // Nonetheless, let's see if performance is OK with this if statement--the compiler may give it to us for free.
-  if (!iter->is_iterator_alive()) { return *this; } // Iterator will be released if there is an error
-
-  auto error = iter->finish_iterator_child();
-  if (error) { return *this; }
-
-  bool has_value;
-  error = iter->get_iterator().has_next_element().get(has_value);
-  if (error) { return *this; }
-  if (!has_value) { iter->finish_iterator(); }
+simdjson_really_inline array_iterator &array_iterator::operator++() noexcept {
+  error_code error;
+  // PERF NOTE this is a safety rail ... users should exit loops as soon as they receive an error, so we'll never get here.
+  // However, it does not seem to make a perf difference, so we add it out of an abundance of caution.
+  if ((error = iter.error()) ) { return *this; }
+  if ((error = iter.finish_child() )) { return *this; }
+  if ((error = iter.has_next_element().error() )) { return *this; }
   return *this;
 }
 
@@ -48,36 +32,30 @@ simdjson_really_inline array_iterator<T> &array_iterator<T>::operator++() noexce
 
 namespace simdjson {
 
-template<typename T>
-simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>::simdjson_result(
-  SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T> &&value
+simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>::simdjson_result(
+  SIMDJSON_IMPLEMENTATION::ondemand::array_iterator &&value
 ) noexcept
-  : SIMDJSON_IMPLEMENTATION::implementation_simdjson_result_base<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>(std::forward<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>(value))
+  : SIMDJSON_IMPLEMENTATION::implementation_simdjson_result_base<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>(std::forward<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>(value))
 {
 }
-template<typename T>
-simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>::simdjson_result(error_code error) noexcept
-  : SIMDJSON_IMPLEMENTATION::implementation_simdjson_result_base<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>({}, error)
+simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>::simdjson_result(error_code error) noexcept
+  : SIMDJSON_IMPLEMENTATION::implementation_simdjson_result_base<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>({}, error)
 {
 }
 
-template<typename T>
-simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::value> simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>::operator*() noexcept {
+simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::value> simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>::operator*() noexcept {
   if (this->error()) { this->second = SUCCESS; return this->error(); }
   return *this->first;
 }
-template<typename T>
-simdjson_really_inline bool simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>::operator==(const simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>> &other) const noexcept {
+simdjson_really_inline bool simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>::operator==(const simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator> &other) const noexcept {
   if (this->error()) { return true; }
   return this->first == other.first;
 }
-template<typename T>
-simdjson_really_inline bool simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>::operator!=(const simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>> &other) const noexcept {
+simdjson_really_inline bool simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>::operator!=(const simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator> &other) const noexcept {
   if (this->error()) { return false; }
   return this->first != other.first;
 }
-template<typename T>
-simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>> &simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator<T>>::operator++() noexcept {
+simdjson_really_inline simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator> &simdjson_result<SIMDJSON_IMPLEMENTATION::ondemand::array_iterator>::operator++() noexcept {
   if (this->error()) { return *this; }
   ++(this->first);
   return *this;