… benchmark/distinctuserid/dom.h
… benchmark/distinctuserid/ondemand.h
… benchmark/find_tweet/dom.h
… benchmark/find_tweet/find_tweet.h
… benchmark/find_tweet/ondemand.h
… benchmark/partial_tweets/ondemand.h
… doc/ondemand.md
… include/simdjson/generic/ondemand/document.h
… include/simdjson/generic/ondemand/field-inl.h
… include/simdjson/generic/ondemand/field.h
… include/simdjson/generic/ondemand/json_iterator-inl.h
… include/simdjson/generic/ondemand/json_iterator.h
… include/simdjson/generic/ondemand/logger-inl.h
… include/simdjson/generic/ondemand/logger.h
… include/simdjson/generic/ondemand/object-inl.h
… include/simdjson/generic/ondemand/object.h
… include/simdjson/generic/ondemand/object_iterator-inl.h
… include/simdjson/generic/ondemand/object_iterator.h
… include/simdjson/generic/ondemand/raw_json_string-inl.h
… include/simdjson/generic/ondemand/raw_json_string.h
… include/simdjson/generic/ondemand/token_iterator.h
… include/simdjson/generic/ondemand/value-inl.h
… include/simdjson/generic/ondemand/value.h
… include/simdjson/generic/ondemand/value_iterator-inl.h
… include/simdjson/generic/ondemand/value_iterator.h
… tests/ondemand/ondemand_basictests.cpp
… tests/test_macros.h