@lemire 2020-12-05
Sounds really good. Marking for 0.8.0.