@jkeiser 2020-12-07
Note: while the performance boost appears to be real, the *size* of the performance boost depends significantly on the size/layout of the benchmark executable. Adding unrelated benchmarks caused the different to shrink from 6% to 2% (and the final version to run at 3.4GB/s instead of 3.6).