@jkeiser 2020-12-07
Regardless, I've added the tests that needed to be added (and found a bug!), and made `object["a"]["b"]` work. Working through multiplatform issues now, then it'll be ready for review.