@jkeiser 2020-12-07
@lemire note document_stream_tests failed with some kind of memory dump--not that it's impossible my changes caused an issue, but they are pretty far away from that code in general ... https://cloud.drone.io/simdjson/simdjson/7343/10/2