@jkeiser 2020-12-07
All tests are passing, and we are still a bit faster (it was even faster in earlier tests, and I think is strongly affected by binary layout).  It has solid tests, and updated documentation. I'll post more numbers soon, but this is ready for review.