@jkeiser 2020-12-07
@lemire will fix :) Sounds like I must have introduced some in the doc update.