@jkeiser 2020-12-07
@lemire your comment about the pretty reminded me of a few other small cleanups I wanted to make after this change. [This commit(https://github.com/simdjson/simdjson/pull/1330/commits/806cb39103c0b0d2f1e9d77fc5914f88131ff425) shows them (nothing big, just quality of life).