@lemire 2020-12-09
@jkeiser This PR looks good. It is always best for me to actually work with the code to judge it properly, but I think you can merge it at will.