@jkeiser 2020-12-09
Ok! I am working through the next PR in my head: I want to restore the ability to validate that you are using it right.